+++ src/utilities/index.ts
@@ -1,4 +1,3 @@
-export * from "./HotspotBuilder";
 export * from "./deepEqual";
 export * from "./mixColor";
 export * from "./typeGuards";